+++ src/Attributes/Collections/AbstractCollectionTransformer.php
@@ -10,6 +10,8 @@
 namespace Cline\Struct\Attributes\Collections;
 
 use Cline\Struct\Contracts\TransformsCollectionValueInterface;
+use Cline\Struct\Contracts\TransformsLaravelCollectionValueInterface;
+use Illuminate\Support\Collection;
 
 /**
  * Base attribute for deterministic collection transforms.
@@ -17,10 +19,15 @@ use Cline\Struct\Contracts\TransformsCollectionValueInterface;
  * @author Brian Faust <brian@cline.sh>
  * @psalm-immutable
  */
-abstract readonly class AbstractCollectionTransformer implements TransformsCollectionValueInterface
+abstract readonly class AbstractCollectionTransformer implements TransformsCollectionValueInterface, TransformsLaravelCollectionValueInterface
 {
     public function supportsLists(): bool
     {
         return true;
+    }
+
+    public function transformCollection(Collection $items): Collection
+    {
+        return new Collection($this->transform($items->all()));
     }
 }