+++ tests/Fixtures/Data/InferredLaravelCollectionData.php
@@ -1,0 +1,27 @@
+<?php declare(strict_types=1);
+
+/**
+ * Copyright (C) Brian Faust
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Tests\Fixtures\Data;
+
+use Cline\Struct\AbstractData;
+use Cline\Struct\Attributes\AsCollection;
+use Cline\Struct\Enums\DataListType;
+use Illuminate\Support\Collection;
+
+/**
+ * @author Brian Faust <brian@cline.sh>
+ * @psalm-immutable
+ */
+final readonly class InferredLaravelCollectionData extends AbstractData
+{
+    public function __construct(
+        #[AsCollection(DataListType::Int)]
+        public Collection $scores,
+    ) {}
+}